Add files via upload
The program employs absorption and scattering probabilities at every thickness step. The path of each electron is tracked as the electron passes through the sample suffering multiple scattering events. The tracking is done both in real space as well as in momentum space. The real space position of the electron is recorded to track the position of the electron near the interface and to determine if it is in the low or high density material. This is important as different scattering probabilities are used for both materials. The program also tracks the angle at which the electron traverses the materials. This corresponds to the x-y components of the electrons momentum vector and to the scattering angle of electrons in the diffraction plane.